feat(#646): Use next due date when selecting a new one
Before, when selecting a due date for an item, we used to pre-select the
original due date of an item. For items with a recurrence pattern, the
first due date can be weeks or months before the current one - which is
quite confusing when you want to select a new due date.

We change this so that the editor now shows the current due date of the
item, which is way more intuitive.