fix(#647): Re-selecting the Original Due Date
We fixed a bug which manifested when one tried to re-set the original due
date on an item with a recurrence pattern. Let's say, if an item is set
to be done the next day initially, with a recurrence pattern of daily
and the user marks the first occurrence as done.

Then the item will be re-scheduled to the same day but a week later.
Now, editing the item and trying to set the original due date again on
it will have no effect! 😱

This is fixed now, so that action will work as expected.