feat(#561): Update images in Play Store and App Store
We updated the screenshots in the Apple's App Store and Google's Play
Store to give a more accurate impression what the users can expect when
installing the app. 🤓

For this, we also applied some minor fixes here and there, for example,
we added an option to create a sample library presenting some of the
features the app supports as well as some improvements to the newly
introduced menus of the app.