doc(#545): Adding Some Images To The README
The README - which is the initial entry point and first thing you see
about the app if you visit it on GitLab or GitHub - was, until now, a
pure text pased thing.

We changed this now by adding some representative screenshots showing
some features of the app. This should give a way better impression about
what the app is about and what new users can expect once installing it.