fix(#642): Disable Backup on iOS
One of the promises the app makes: No data ever leaves the device you
work on (except you explicitly set up e.g. a library which is
synchronized via one of the supported storage services). This is true on
all platforms, but one: on iOS, your libraries actually made it into the
iCloud backup (if you had it set up) 😱

This is not only a problem of control of your data - it also can create
some friction when e.g. you move to a new device. It seems nice that
your libraries appear to be back automagically, but because of the
timestamps of the local files are changes, the new device will basically
re-upload the entire library in case it is also synced via a storage
server. Long story short: This is fixed now!