feat(#637): Automate Uploading To Google Play
Automation - especially of very regularly repeating tasks - is key to
optimize for efficiency and gain some time for the more interesting
things in life - such as developing new features in your favourite todo
app 😉

Hence, in this release we finally automated uploading of Android releases
to the Play Store! Before, this was a manual step: After a release
pipeline ran, we had to download the AAB files from the pipeline manually
and upload them again in the Play Console - a boring and time consuming
job. That - now - is automated; in the future, we can do the same with a
push on a button.