feat(#554): Links to Libraries and Items
Creating and passing around links to certain things is a quite useful
feature - the entire WWW lives around this, after all. OpenTodoList now
allows for the same: Links to libraries as well as items 🎉

So, how does this work? You can now navigate to a library or item and in
the edit menu, select the "Copy Link To Page" option. This will copy a
special link to the clipboard - such links start with the "opentodolist"
scheme instead or "https" as usual web links. You can then insert such
links into item titles or notes and later on click on them - this will
directly open the linked to item in the current page stack.

But, it doesn't stop there 😉 On some systems, the app also will be
opened if another application opens such a link. For example, if you
have a library together with someone else, you can send them a link to
one of the items via chat and - if everything is set up accordingly -
the app should open up and show that item. Neat, isn't it?

Currently, this is supported on iOS and Android, further operating
systems will follow in the future.